Refactor JPEG2000 plugin dispatch internals
